// src/services/SignUpService/types.ts
export enum SignUpPage {
  account = 'account',
  eligibility = 'eligibility',
  ineligible = 'ineligible',
  international = 'international',
  parentGuardianConfirmation = 'confirmation',
  partnerInfo = 'info',
  verify = 'verify',
  about = 'about',
}

export enum UserType {
  student = 'student',
  teacher = 'teacher',
  volunteer = 'volunteer',
}